Jormungandr: Cast datetime in request to integer
